requirements.txt
uvicorn==0.34.0
fastapi==0.115.11
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
pydub==0.25.1
openai==1.66.3
azure-monitor-opentelemetry == 1.6.4